+++ R/later.R
@@ -87,11 +87,23 @@ later <- function(func, delay = 0) {
 #' or control returns to the R prompt). You must use your own
 #' [tryCatch][base::conditions] if you want to handle errors.
 #'
+#' @param timeoutSecs Wait (block) for up to this number of seconds waiting for
+#'   an operation to be ready to run. If `0`, then return immediately if there
+#'   are no operations that are ready to run. If `Inf` or negative, then wait as
+#'   long as it takes (if none are scheduled, then this will block forever).
+#'
 #' @return A logical indicating whether any callbacks were actually run.
 #'
 #' @export
-run_now <- function() {
-  invisible(execCallbacks())
+run_now <- function(timeoutSecs = 0L) {
+  if (timeoutSecs == Inf) {
+    timeoutSecs <- -1
+  }
+  
+  if (!is.numeric(timeoutSecs))
+    stop("timeoutSecs must be numeric")
+  
+  invisible(execCallbacks(timeoutSecs))
 }
 
 #' Check if later loop is empty