+++ src/later.h
@@ -1,7 +1,7 @@
 #ifndef _LATER_H_
 #define _LATER_H_
 
-bool execCallbacks();
+bool execCallbacks(double timeoutSecs = 0);
 bool at_top_level();
 bool idle();
 extern "C" void execLaterNative(void (*func)(void*), void* data, double secs);